+++ src/node.rs
@@ -39,3 +39,27 @@ pub(crate) fn children_of<T>(nodes: &Ref<Vec<InnerNode<T>>>, id: &NodeId) -> Vec
     }
     children
 }
+
+pub(crate) fn ancestors_of<T>(nodes: &Ref<Vec<InnerNode<T>>>, id: &NodeId, max_depth: Option<usize>) -> Vec<NodeId> {
+
+    let max_depth= max_depth.unwrap_or(0);
+    let mut depth: usize = 0;
+
+    let mut ancestors = vec![];
+
+    if let Some(node) = nodes.get(id.value) {
+        let mut parent = node.parent;
+        while let Some(parent_id) = parent {
+            if max_depth > 0 && depth == max_depth {
+                break;
+            }
+
+            ancestors.push(parent_id);
+            depth += 1;
+            if let Some(node) = nodes.get(parent_id.value) {
+                parent = node.parent;
+            }
+        }
+    }
+    ancestors
+}